arm64: dts: Temporarily disable griffin 1096p res
Currently griffin's 4K panel is having a really hard time due to Android
userland not being able to correctly figure out what is happening: the
resolution switcher would crop the screen when switching from 1096p to 4K.

On the other hand, inverting the values of timing@0 and timing@1 would
result in the screen having massive black borders when in 1096p mode (as
found out by Re:Shion).

Thus, disable the 1096p resolution in the meantime while we explore the
possibility of even having proper resolution switching on an sm8150 device.

Link: https://xdaforums.com/t/rom-unofficial-lineageos-20-21-for-xperia-1-5.4630585/page-12#post-89381718
Signed-off-by: voidanix <voidanix@keyedlimepie.org>